Fix add service & stuff and search & filter reservation

+++ app/DAO/Deposit_DAO.cs
@@ -49,7 +49,7 @@ namespace app.DAO
             Deposit_DTO deposit = new Deposit_DTO();
             foreach(DataRow item in table.Rows)
             {
-                deposit.Deposit = (double)item["deposit"];
+                deposit.Deposit = (double)((decimal)item["deposit"]);
                 break;
             }
             return deposit.Deposit;
